[IMP] mail: add tool methods to manage activities directly on the mixin
This commit adds tools methods related to activities in the mixin. It
gives to models inheriting from the activity mixin an easy-to-use API to
schedule, unlink or activities as done. Purpose of those methods is to
avoid having people manually managing activities in the code to hide
the technical details of the activities model, notably access rights
or activity types.

A field is added on activity model to indicate they have been generated
automatically. This way when rescheduling or unlinking based on some
specific activity types we do not change user-created activities.

Those methods include scheduling activities, changing their dates, marking
them as done or unlinking them. Future commits will use those methods
in various addons to automatically generate activities based on workflow
we want to implement.

This commit also adds tests for the newly added code.