[IMP] mail, hr_recruitment: add a mail mixin for partner creation.
This mixin aims to replace some redundant logic in some applications and to
autosubscribe a partner when replying on a record by email.
The main problem is linked to the fact that when a record is created by email
the partner is not added to the followers. He will never get the answer. Adding
this mixin will solve the problem by creating a partner if it doesn't exists and
subscibing it when needed. Moreover redundant logic, like linking the new partner
to a specific field of the record have been added to this mixin.

Task #46151
PR #22608