[IMP] base: prevent to archive users linked to active partners.
An active user could not see messages in his inbox if his partner is
archived. With this commit, we cannot archiving a partner linked to a
user
anymore.

In the opposite way, if we deactivate a user, warning message is given
saying the related partner will stay active. A link to the related
partner is also given if we really need to archive it.