[IMP] sale_timesheet: tasks linked to SOL of project
When selling a project from a sale order, the SO line
creates the project. When we create a task in that
kind of project, by default, the task will be linked
to the SO line of its project. This can still be
changed manually (no constraint).
So, every time spend on the project (each timesheet on
every tasks) are counted into the delivered quantity
of the SO line.