[IMP] mail: add code on message subtypes allowing to lessen master data effect
Purpose of this commit is to lessen dependency on some xml ids in various
place of code related to the mail module. Indeed mail.mt_comment,
mail.mt_note and mail.mt_activities xml ids are commonly used to fetch
the discussion, note and activities subtypes.

This causes several issues. If people delete those subtypes the whole
discussion mechanism is going to be completely unusable, including Chatter
or the mail gateway. It is not possible to reconfigure them without doing
complex technical manipulations in order to make the link between the
subtypes and the xml ids.

This commit proposes to solve that issue by adding a code field on
subtypes. Using the name is not feasible as it is a translatable field
that people could also change. Code is a technical field that holds the
specific meaning of that subtype.

An override of write and unlink is added to ensure there is always
at last one subtype of each magic kind, currently note, discussion and
activities. That way people cannot break their Odoo but the dependency
is softened compared to xml id dependencies.

In this commit we also provide some cached methods allowing to keep
specific subtypes ids in cache. We hope its use in various place of code
will allow to gain some queries.