[REM] mail: remove o_mail_note class in chatter
Purpose of this commit is to clean a bit various classes and conditions.

First clean is the removal the class in activities less. It comes from
first implementation of activity widget and is not used anymore.

Second clean is the removal of the class when displaying messages. Only
o_mail_discussion is used currently in chatter. All we have to know to
differentiate message display is whether it is a discussion done by an
user or not.

Tests have been updated to use o_mail_discussion class triggered by
the message being a discussion, aka a comment posted by an user.