[IMP] stock: Picking Operation report set printed
Before this commit there was 2 ways to print the
picking operation report. Classic way with the print dropdown.
Second on the fom view with a button 'print' in the header buttons.

However if the user uses the 'print' button on the form view
it will trigger a server action that will set the variable printed
to true on the picking. That was not the case if the user uses the
classic way to print report with the dropdown menu.

The header button 'print' returns 2 differents report depending
the picking state. Delivery slip if the picking is done else the
picking operations. This two reports should be always available no
matters the picking state for these reasons we decide to remove the
print button from the header.