[IMP] base_vat: reintroduce colors to visualize VAT number verification status
[IMP] base_vat: no stored computed field for VIES check status anymore

Using a stored computed field the way we used to was annoying for multi company. Indeed, we could not make it company dependent (as we don't mind the current company when checking VAT numbers, if it's correct for one, it's correct for all). Furthermore, making it a regular computed field would have meant the service would be called even for companies that did not activate the VIES option, just because some other company did install the module, resulting in useless operations.

Instead, we use a non-stored computed field to store the VAT check status (offline or VIES), which will take its value depending on the curreny company's options. For VIES, we use a regular Selection field hidden to the user, into which we cache the result of the call to the webservice.