[IMP] web : manage up and down keys in list view
before, the up and down keys were not used in list view, and the default focus was on the search view
now, when opening a list view, the default focus is still on the search view. Hitting down will select the first
record of the list view. From there, using UP and DOWN with move the selection from record to record in the list
view.