[IMP] sale_timesheet: wizard to create SO from project
This commit provides a wizard to create a SO from
a non billable project. It set sale_line_id and
others fields like the project was created on SO
confirmation.
The created SO contains one line with a selected
service product, with ordered qty set to zero.
The wizard also links all non billable active
tasks of the project to the SO line (and its
timesheet entries).