[FIX] models: do not check Python constraints on partial updates
Consider a model with a constraint method that depends on a stored field
`country_id` and a non-stored computed field `currency_id`:

    @api.constrains('country_id', 'currency_id')
    def _check_country_currency(self):
        # check whether country_id and currency_id match each other

When writing on both `country_id` and `currency_id`, the method is invoked
twice: the first time when `country_id` is updated, and the second time where
both fields are updated.  In this case, the first invocation may fail, while
the second does not.  Overall it fails.

Fix this by not checking constraints on partial updates.

OPW 1817493