[FIX] website_sale{,*}: update when changing quantity
On the product page or "Add to Cart" modal, the price of the product was
never changed.

This happened seemingly since 65e810269 because after this, we get
invalid JSON in the attribute_value_ids attribute.

This cause the .data(...) to sends back a string that we then parse. But
then the code is expecting to modify it and have it updated.

This change fix the original issue by doing `json.dumps` when adding the
attribute_value_ids attribute, thus we get an object directly.

For instance that would not update views (in website_sale and
website_sale_options), there is a second fix that update
.data('attribute_value_ids') with the reference to the parsed object.

opw-1816479
closes #23392