[REF] mail: use new followers command in message_subscribe
UPDATE ME

This commit also changes default behavior of message_subscribe. UPDATE ME

This commit has a great impact when auto subscription is involved. Real
life use case is creating a task in a project with several followers.
Using the new followers subscription allow to save a lot of queries.