[FIX] stock: delivery slip report and ordered quantities
The delivery report was displaying a lot of informations using the
`ordered_qty` field on the move line. Sadly, this field just doesn't
work as it should: it represents the first reserved quantity and is
never updated. The goal of the real `ordered_qty` field on the move is
to know how much was ordered from the SO, to know if we deliver more or
not. We cannot adapt this field on the move line.

Also, these informations should not be displayed on the delivery slip.
This document is desinated to a customer. We thus remove everything a
custome should not care about in the delivery slip: lot, delivered vs
ordered quantities (there's a text at the end indicating if a backorder
delivery will follow).

opw-1808126