[POC] pos: display warning when a session lasts too long
A reucrring problem with our POS system is big sessions with many
orders.

Our documentation mentions in passing that you should close your
sessions each day, but nothing else does and this is not obvious
at all that you will run into problems if you don't.

This is a vicious problem because when a session becomes too big,
you won't be able to validate and close it in your normal usage
and will either require your partner's or our intervention, typically
with invasive operations such as lowering timeout times or launching
separate processes that close the session 'on the side'.

This commit introduces a new 'warning' system that advises the user to
close its sessions regularly using a yellow warning after 2 days of
session use then a red warning after 5 days on the same session in the
status bar of the pos interface (near the 'close' button).

This does not in itself prevent the user from having huge sessions if
the POS is heavily used in a short period, but it should at least make
customers aware the limits of our system in advance rather than too
late.