[FIX] barcode: don't bypass _barcode_scanned on x2many
There is a small optimisation on _barcodeAddX2MQuantity when
a candidate is find, the system will forge the onchnage in order
to only update the selected record in the X2Many. However barcode
mixin is based on an onchange of the field _barcodes_scanned which
is bypass in this case.

That could have some consequence since some condition/actions are
server side.