[IMP] project: set project_id mandatory on tasks
This commit forces a task to belong to a project. For
historical reason, the project_id field on project.task
is optional. In 2018, it seems that it makes no sense
anymore. Moreover, this might create some problems
for further developpement:
- This makes task without project difficult to find
- When a task have some timesheets, transfering the
task from a project to another is difficult but possible.
Removing the project of this kind of task is impossible,
as a analytic line need a project to be considered as a
timesheet.
- Without a project, it is also impossible to forecast
employee time.

So, we don't have any interesting use cases to keep this
field mandatory.

However, this break the quick create of a task, as there
is no default value (coming from Project > Search > Tasks).
The solution is to display project field on quick create
view when there is no default value.

Also, with this modification, this commit simplifies some
useless checks, since project is now required.

Tasks #34703